# improve test_custom_function