10 to 20x faster 

now:

```
CPU=1 python3 test/test_custom_function.py 
test_atan2_backward tg: 8.4061 ms, np: 2.7502 ms
...
----------------------------------------------------------------------
Ran 3 tests in 0.055s
```

before:

```
CPU=1 python3 test/test_custom_function.py 
[[-0.06867199 -0.00292654  0.00270511  0.01674124]
 [ 0.00136377  0.0646913   0.02751377 -0.07913955]
 [ 0.02835985  0.14661054 -0.05507449  0.01771379]
 [-0.01024144 -0.03046751 -0.00165082  0.03411034]]
[[-0.01191162 -0.08596582 -0.07924119 -0.02560918]
 [-0.04759943 -0.11794868  0.01067733  0.0020435 ]
 [ 0.02315945  0.04482204  0.05077991 -0.0404908 ]
 [-0.03755711 -0.05102779  0.10992462  0.00390302]]
.[[-1.256934    2.8877442   1.8022928   2.8673573 ]
 [ 1.5012058   1.8774074   0.02821589 -2.3339114 ]
 [ 2.4752936  -1.0364021   2.63329    -1.275166  ]
 [ 0.9826964  -2.3509507   2.885475    2.9024506 ]]
..
----------------------------------------------------------------------
Ran 3 tests in 0.997s
```