@geohot 2023-07-18
This is not worth it, 1s is so fast and I like that torch computes the gradient itself.