@foreign-sub 2023-07-18
sure 1s is fast, but those 1s's add up pretty fast when you run those tests a lot, especially if you want to run ci in less than 5 minutes.

anyway it's there if you need it.